+++ test/metaevent.jl
@@ -41,4 +41,12 @@
         sse = MIDI.SequencerSpecificEvent(0, 0x7f, [0x11, 0x21, 0x53, 0x1F])
         @test sse.ssdata == [0x11, 0x21, 0x53, 0x1F]
     end
+
+    @testset "check empty" begin
+        midi = load(testmidi())
+        track = midi.tracks[1]
+        @test !isempty(track)
+        empty!(track)
+        @test isempty(track)
+    end
 end