+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 # v2.7.0
 * New meta event type `SequencerSpecificEvent`.
-* Implement `empty!(::MIDITrack)`.
+* Implement `Base.empty!(::MIDITrack)` and `Base.isempty(t::MIDITrack)`.
 # v2.6.0
 * New functions `metric_time`.
 * New functions `duration_metric_time`.